backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Environment variables
python-decouple==3.8

# API & HTTP
requests==2.31.0
httpx==0.26.0

# Async support
celery==5.3.4
redis==5.0.1

# Data validation
pydantic==2.5.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
django-debug-toolbar==4.2.0
ipython==8.19.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
